refactor: extract ContentType, HttpStatusCode, and header types into separate modules
