+++ src/rpc/lib/http-status-code-types.ts
@@ -1,0 +1,93 @@
+/**
+ * Informational responses (100–199)
+ */
+type InformationalHttpStatusCode = 100 | 101 | 102 | 103;
+
+/**
+ * Successful responses (200–299)
+ */
+export type SuccessfulHttpStatusCode =
+  | 200
+  | 201
+  | 202
+  | 203
+  | 204
+  | 205
+  | 206
+  | 207
+  | 208
+  | 226;
+
+/**
+ * Redirection messages (300–399)
+ */
+export type RedirectionHttpStatusCode =
+  | 300
+  | 301
+  | 302
+  | 303
+  | 304
+  | 305
+  | 306
+  | 307
+  | 308;
+
+/**
+ * Client error responses (400–499)
+ */
+type ClientErrorHttpStatusCode =
+  | 400
+  | 401
+  | 402
+  | 403
+  | 404
+  | 405
+  | 406
+  | 407
+  | 408
+  | 409
+  | 410
+  | 411
+  | 412
+  | 413
+  | 414
+  | 415
+  | 416
+  | 417
+  | 418
+  | 421
+  | 422
+  | 423
+  | 424
+  | 425
+  | 426
+  | 428
+  | 429
+  | 431
+  | 451;
+
+/**
+ * Server error responses (500–599)
+ */
+type ServerErrorHttpStatusCode =
+  | 500
+  | 501
+  | 502
+  | 503
+  | 504
+  | 505
+  | 506
+  | 507
+  | 508
+  | 510
+  | 511;
+
+/**
+ * Http status code (100～599)
+ */
+export type HttpStatusCode =
+  | InformationalHttpStatusCode
+  | SuccessfulHttpStatusCode
+  | RedirectionHttpStatusCode
+  | ClientErrorHttpStatusCode
+  | ServerErrorHttpStatusCode;
