+++ src/rpc/server/index.ts
@@ -1,2 +1,4 @@
 export { routeHandlerFactory } from "./route-handler-factory";
-export type { ContentType, HttpStatusCode, TypedNextResponse } from "./types";
+export type { TypedNextResponse } from "./types";
+export type { ContentType } from "../lib/content-type-types";
+export type { HttpStatusCode } from "../lib/http-status-code-types";
